feat: create an error class extension from the original error class
Why did i do this? because i had no way to return the error from the service to the controller being able to specify the status code of the error making the controller specify the status code by comparing the returned message with a list of messages
what's bad about that? the thing is that if i wanted to change the error message in the service i'd also have to change the corresponding error message in the controller so it'd be identified

- learn how to create an extended class in typescript